+++ CHANGES.md
@@ -5,8 +5,10 @@
 ### Added
 
 + core: Add `AsyncComponentStream` to supports streams for async components as well
++ components: Add `SimpleComboRow` helper for libadwaita's `ComboRow`, analogous to `SimpleComboBox`
++ components: Add an example for `SimpleComboRow`
 + components: Add an example how to use `SimpleComboBox`
- 
+
 ### Changed
 
 + core: Just require `AsRef<gtk::Window>` in `RelmApp::run`